+++ source/ui/controls/spinnerControl.cpp
@@ -20,7 +20,7 @@ void uiSpinner(const char* label, const float indicator_radius, const ImVec4& ma
     if (!ItemAdd(bb, id)) {
         return;
     }
-    const float t = g.Time;
+    const float t = (float)g.Time;
     const auto degree_offset = 2.0f * IM_PI / circle_count;
     for (int i = 0; i < circle_count; ++i) {
         const auto x = indicator_radius * std::sin(degree_offset * i);
